Initial fake juju CLI skeleton (#4)
Add a new fake-juju command and associated Go source and make
target. This command will behave exactly like the regular 'juju' cli
tool, except that it will always point to a fake-jujud instance.

This branch adds the initial code to write files to the JUJU_DATA
directory, pointing to a fake-jujud controller.

Integration tests have been added to exercise this initial behavior,
and assert that the expected files are actually written in the JUJU_DATA
directory.