Runner main loop (#6)
Make the main loop of the fake-jujud service a bit more sophisticated

Essentially, instead of unconditionally kick off juju's testing
machinery (i.e. the JujuConnSuite that FakeJujuSuite inherits from,
and which basically bootstraps the dummy provider/controller), it will
enter a "control loop", waiting for commands. This allows a better
control of the service flow, where the dummy controller can be
bootstrapped and destroyed at fake-jujud runtime, without restarting
the service.

The actual commands implementation and a simple HTTP API to kick them
off will be pushed in follow-up branches.