Initial plumbing for fake-jujud's control plane API (#7)
This branch adds some plumbing code that makes the fake-jujud process
start a simple HTTP server.

No request handlers have been defined yet, but the plan is to add them
in follow-up branches. They will offer a "control plane" API which is
specific to fake-juju, such as:

- perform a fake bootstrap using the dummy provider
- tell fake-jujud that certain units or machines should fail

etc.

The branch also adds a new hunk to juju-core.patch, to support
specifying the port that juju's testing machinery (the one that
fake-juju is built on) will use when it starts the test API
server. This allows fake-jujud to have predictable ports for both the
API server and the control plane API.

By default, the API server port will be 17099 and the control plane
API port will be 17100, but that can be changed with the newly
introduced "-port" command line switch. Note that the control plane
API port is always equal to the API server port plus one.