Add more development-oriented documentation (#9)
This branch adds a new DESIGN.txt file holding basic information about fake-juju's design and code organization. It should make it easier for people not familiar with the code (or who haven't looked at it for a while) to get a grasp on it.

It also moves some development-related documentation from README.rst to HACKING.rst, complenting it with missing bits, such else creating a new quilt patch.