Implement the core delta stream watch loop (#19)
This branch adds the core bit of fake-juju's behavior: a for loop that
will watch the websocket delta stream and react appropriately when
entities are added or removed (e.g. automatically start units and
machines).

Note that only the core loop logic and related error/exit mechanisms
are implemmented here. The actual logic for handling specific entity
changes will be added in follow-up branches.