Automatically transition machines to the "started" state (#20)
The main goal of this branch is to introduce logic that reacts to new
machine deltas coming from the core fake-jujud watch loop, and
automatically transition machines from the "pending" to the "started"
state.

The entry point of this logic is the new handleMachineChanged method
of the new machine.go file, which is what will be invoked by the main
watch loop.

Note that some code (basically the startMachine method) has been moved
from service.go to machine.go, as the former was starting to get a bit
big and more code for handling other kinds of entities will come in
follow-up branches.

Lastly, the InitializeController() code has been dropped altogether,
since it was basically special-casing machine 0 and starting it "by
hand". Now, machine 0 will be started automatically via
handleMachineChanged, like any other machines.