Don't automatically start new machines for unbound units (#30)
This branch changes the default behavior of fake-juju to *not*
automatically start machines for new units that don't appear to have
one yet.

This behavior must now be explicitly enabled with a flag, which at the
moment is only used by unit tests, for convenience.

Initially I added this flag for human/test convenience, since the
Landscape autopilot rigorously waits for machines to start before
bounding units to them. However, the delta watcher seems to sometime
notify new units *without* including the machine they are bound to,
even if that machine exists at the time that add-unit was invoked.